Moved ConfigManager.java to ..statisticslib.plugin
