+++ docs/org/dreamndelight/playerstatistics/lib/data/ConfigManager.html
@@ -10,20 +10,19 @@
 </head>
 <body>
 <script type="text/javascript"><!--
-    try {
-        if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="ConfigManager";
-        }
+try {
+    if (location.href.indexOf('is-external=true') === -1) {
+        parent.document.title = "ConfigManager";
     }
-    catch(err) {
-    }
+} catch (err) {
+}
 //-->
-var methods = {"i0":10,"i1":10,"i2":10,"i3":10};
-var tabs = {65535:["t0","All Methods"],2:["t2","Instance Methods"],8:["t4","Concrete Methods"]};
-var altColor = "altColor";
-var rowColor = "rowColor";
-var tableTab = "tableTab";
-var activeTableTab = "activeTableTab";
+const methods = {"i0": 10, "i1": 10, "i2": 10, "i3": 10};
+const tabs = {65535: ["t0", "All Methods"], 2: ["t2", "Instance Methods"], 8: ["t4", "Concrete Methods"]};
+const altColor = "altColor";
+const rowColor = "rowColor";
+const tableTab = "tableTab";
+const activeTableTab = "activeTableTab";
 
 </script>
 <noscript>
@@ -63,15 +62,14 @@ var activeTableTab = "activeTableTab";
         <li><a href="../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
     </ul>
     <div>
-        <script type="text/javascript"><!--
-  allClassesLink = document.getElementById("allclasses_navbar_top");
-  if(window==top) {
-    allClassesLink.style.display = "block";
-  }
-  else {
-    allClassesLink.style.display = "none";
-  }
-  //-->
+        <script type="text/javascript">
+            allClassesLink = document.getElementById("allclasses_navbar_top");
+            if (window === top) {
+                allClassesLink.style.display = "block";
+            } else {
+                allClassesLink.style.display = "none";
+            }
+            //-->
 
         </script>
     </div>
@@ -104,7 +102,7 @@ var activeTableTab = "activeTableTab";
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>com.devflask.statisticslib.lib.data.ConfigManager</li>
+                <li>com.devflask.statisticslib.plugin.ConfigManager</li>
             </ul>
         </li>
     </ul>
@@ -371,15 +369,14 @@ extends java.lang.Object</pre>
         <li><a href="../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
     </ul>
     <div>
-        <script type="text/javascript"><!--
-  allClassesLink = document.getElementById("allclasses_navbar_bottom");
-  if(window==top) {
-    allClassesLink.style.display = "block";
-  }
-  else {
-    allClassesLink.style.display = "none";
-  }
-  //-->
+        <script type="text/javascript">
+            allClassesLink = document.getElementById("allclasses_navbar_bottom");
+            if (window === top) {
+                allClassesLink.style.display = "block";
+            } else {
+                allClassesLink.style.display = "none";
+            }
+            //-->
 
         </script>
     </div>
